fix warnings

+++ virtual-machine/src/interpreter.cpp
@@ -919,21 +919,29 @@ void Interpreter::ExecuteJitFunction(RuntimeFunction* function, CallFrame& calle
                 case 0: {
                     c = {
                         type_kind.kind == TypeDesc::Kind::BOOL ? ConstantTag::BOOL : type_kind.is_signed ? ConstantTag::I1 : ConstantTag::U1, 
-                        {stack[return_index] & 0xFF}
+                        {static_cast<uint8_t>(stack[return_index] & 0xFF)}
                     };
                     break;
                 }
                 case 2: {
                     c = {
                         type_kind.is_signed ? ConstantTag::I2 : ConstantTag::U2, 
-                        {stack[return_index] >> 8, stack[return_index] & 0xFF}
+                        {
+                            static_cast<uint8_t>(stack[return_index] >> 8), 
+                            static_cast<uint8_t>(stack[return_index] & 0xFF)
+                        }
                     };
                     break;
                 }
                 case 4: {
                     c = {
                         type_kind.is_signed ? ConstantTag::I4 : ConstantTag::U4, 
-                        {stack[return_index] >> 24, (stack[return_index] >> 16) & 0xFF, (stack[return_index] >> 8) & 0xFF, stack[return_index] & 0xFF, }
+                        {
+                            static_cast<uint8_t>(stack[return_index] >> 24),
+                            static_cast<uint8_t>((stack[return_index] >> 16) & 0xFF),
+                            static_cast<uint8_t>((stack[return_index] >> 8) & 0xFF),
+                            static_cast<uint8_t>(stack[return_index] & 0xFF)
+                        }
                     };
                     break;
                 }
